% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/parse.R
\name{parse_results}
\alias{parse_results}
\title{Parse results of a request}
\usage{
parse_results(request, format)
}
\arguments{
\item{request}{}
}
\value{
data.frame of reusults.
}
\description{
Parse results of a request
}
\keyword{internal}

