% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/sdb.r
\name{ratify_access_point}
\alias{ratify_access_point}
\title{validate access_point and retrieve it.}
\usage{
ratify_access_point(access_point)
}
\arguments{
\item{access_point}{(character string)}
}
\value{
access point uri (character string)
}
\description{
validate access_point and retrieve it.
}
\keyword{internal}

