% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/get_results.R
\name{ratify_sdb}
\alias{ratify_sdb}
\title{Validate the structure sdb object}
\usage{
ratify_sdb(sdb)
}
\arguments{
\item{sdb}{sdb}
}
\value{
sdb object (named list)
}
\description{
Validate the structure sdb object
}
\keyword{internal}

