% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/sdb.r
\name{sdb}
\alias{sdb}
\title{Build a STRING-db query}
\usage{
sdb(access_point)
}
\arguments{
\item{access_point}{what to query for (character string). See
details for available options.}
}
\description{
Build a STRING-db query
}

