% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/constrain.R
\name{sdb_constrain}
\alias{sdb_constrain}
\title{contrain STRING-db query with additional parameters}
\usage{
sdb_constrain(sdb, ...)
}
\arguments{
\item{sdb}{sdb object}

\item{...}{additional parameters}
}
\value{
modified bg object. N.B. current behaviour overwites and previously
present params.
}
\description{
contrain STRING-db query with additional parameters
}

