% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/get_results.R
\name{sdb_get_results}
\alias{sdb_get_results}
\title{GET results of a query to the STRING-db web service.}
\usage{
sdb_get_results(sdb, as = "tsv", file = NULL, .request = FALSE)
}
\arguments{
\item{sdb}{sdb object returned by \code{\link{sdb}}. (named list)}

\item{as}{format in which to return results. (character string)}

\item{file}{Specify name of file if you wish to write the results to file.
(character string). N.B. not yet implemented.}
}
\description{
GET results of a query to the STRING-db web service.
}

