% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/zzz.R
\docType{package}
\name{stringdbr}
\alias{stringdbr}
\alias{stringdbr-package}
\title{Convenient and Reproducible Access to STRING-db data}
\description{
Convenient and Reproducible Access to STRING-db data
}
\section{Design}{


Most if not all functions start with \code{sdb_*} for convenient auto-
completion.

Package designed to play with %>% from maggritr.
}

